Added circle-ci auto build

+++ assignment-4/build-wasm.sh
@@ -18,6 +18,10 @@ CONFIG=./assets/config.xml
 
 # Turn frameCap for wasm because wasm already has its own frameCap
 sed -i 's\<frameCapOn>true</frameCapOn>\<frameCapOn>false</frameCapOn>\g' ${CONFIG}
-docker run --rm -v ${PWD}:/src lherman/sdl emcc -o index.html ${FLAGS[@]} ${SRCS[@]}
+if [[ -z ${CI} ]]; then
+    docker run --rm -v ${PWD}:/src lherman/sdl emcc -o index.html ${FLAGS[@]} ${SRCS[@]}
+else
+    emcc -o index.html ${FLAGS[@]} ${SRCS[@]}
+fi
 # Turn the frameCap back on
 sed -i 's\<frameCapOn>false</frameCapOn>\<frameCapOn>true</frameCapOn>\g' ${CONFIG}
